+++ src/Nodes/Operators/LogicalOperators/Sinks/NetworkSinkDescriptor.cpp
@@ -4,12 +4,12 @@ namespace NES {
 namespace Network {
 
 NetworkSinkDescriptor::NetworkSinkDescriptor(NodeLocation nodeLocation, NesPartition nesPartition,
-                                             std::chrono::seconds waitTime, uint8_t retryTimes) :
+                                             std::chrono::seconds waitTime, uint32_t retryTimes) :
     nodeLocation(nodeLocation), nesPartition(nesPartition), waitTime(waitTime), retryTimes(retryTimes) {
 }
 
 SinkDescriptorPtr NetworkSinkDescriptor::create(NodeLocation nodeLocation, NesPartition nesPartition,
-                                                std::chrono::seconds waitTime, uint8_t retryTimes) {
+                                                std::chrono::seconds waitTime, uint32_t retryTimes) {
     return std::make_shared<NetworkSinkDescriptor>(NetworkSinkDescriptor(nodeLocation, nesPartition,
                                                                          waitTime, retryTimes));
 }
